// Projekat/frontend/src/app/models/book.ts
export class Book {
    id: number;
    title: string;
    genre: string;
    language: string;
    publishYear: number;
    publisher: string;
    img: string;
    numOfFreeBooks: number;
    writers: string;
    caption:string;
    bookOfDay:boolean;
    rating: number;
    numOfRatings: number;
    numOfBorrowing: number;
    numOfAllBooks: number;
    username:string;
}